DLF big update

+++ DLF/Optimizers/StochasticGradientDescent.cs
@@ -1,28 +1,36 @@
 using System.Collections.Generic;
 
-namespace DLFramework.Optimizers {
-    public class StochasticGradientDescent {
+namespace DLFramework.Optimizers
+{
+    public class StochasticGradientDescent
+    {
 
-        private float alpha;
+        private double alpha;
         private List<Tensor> parameters;
         public List<Tensor> Parameters { get => parameters; set => parameters = value; }
-        public float Alpha { get => alpha; }
+        public double Alpha { get => alpha; }
 
-        public StochasticGradientDescent (List<Tensor> parameters, float alpha = 0.1f) {
+        public StochasticGradientDescent(List<Tensor> parameters, double alpha = 0.1)
+        {
             this.parameters = parameters;
             this.alpha = alpha;
         }
 
-        public void Zero () {
-            foreach (var parameter in parameters) {
-                parameter.Gradient.Data *= 0f;
+        public void Zero()
+        {
+            foreach (var parameter in parameters)
+            {
+                parameter.Gradient.Data *= 0;
             }
         }
 
-        public void Step (bool zero = true) {
-            foreach (var parameter in parameters) {
+        public void Step(bool zero = true)
+        {
+            foreach (var parameter in parameters)
+            {
                 parameter.Data -= parameter.Gradient.Data * alpha;
-                if (zero) {
+                if (zero)
+                {
                     parameter.Gradient.Data *= 0;
                 }
             }
